% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_profile_evaluation.R
\name{evaluate_gene_profile}
\alias{evaluate_gene_profile}
\title{Compare the abundance matrix from StrainPanDA and ground truth.}
\usage{
evaluate_gene_profile(predicted, ground_truth)
}
\arguments{
\item{x}{Describe the format and meaning of param x}

\item{y}{Describe the format and meaning of param y}
}
\value{
The distance between predicted abundance matrix and ground truth.
}
\description{
Compare the abundance matrix from StrainPanDA and ground truth.
}
\examples{
fun(1, 1)
fun(10, 1)
}
