% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNMFRank.r
\name{findNMFRank}
\alias{findNMFRank}
\title{findNMFRank}
\usage{
findNMFRank(
  obj,
  gene.family.min = 3000,
  presence.thre = 0.5,
  jc.threshold = 0.1,
  abun.threshold = 0.1,
  ncpu = 1,
  max.strain = 12,
  seed = 123456,
  ...
)
}
\arguments{
\item{obj}{strainpandar object (output of `preprocess`)}

\item{gene.family.min}{gene family number used for gene presence detection}

\item{presence.thre}{gene presence threshold}

\item{jc.threshold}{jaccard distance threshold}

\item{abun.threshold}{mean abundance threshold}

\item{ncpu}{number of processes running in parallel (default: 1)}

\item{max.strain}{the maximum number of strains to try}

\item{seed}{used for nmf}

\item{...}{other parameters for nmf}
}
\description{
findNMFRank
}
\author{
Xbiome
}
