% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fprausnitzii.r
\docType{data}
\name{fprausnitzii}
\alias{fprausnitzii}
\title{Demo dataset}
\format{
A list created based on simulated data (randomly drawing reads from WGS data)
\describe{
  \item{data}{Count data created by running PanPhlAn map}
  \item{ans}{Ground truth}
  \item{pangenome}{Data read from the corresponding pangenome file}
  ...
}
}
\usage{
fprausnitzii
}
\description{
Demo dataset
}
\keyword{datasets}
