% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.presence.absence.r
\name{gene.presence}
\alias{gene.presence}
\title{gene.presence}
\usage{
gene.presence(p, n.gene.expected, adjust = 0)
}
\arguments{
\item{p}{one column from the predicted P matrix}

\item{n.gene.expected}{expected number of genes}

\item{adjust}{adjust the threshold between start of the presence peak and the presence peak, must be [0,1]. Smaller value means more relaxed. [Default: 0]}
}
\description{
Determine whether a gene is present or not using a peak finding approach
}
\author{
Xbiome
}
