% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.r
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(
  count.matrix,
  pangenome.file,
  min.cov = 10,
  frac.gene = 0.9,
  scaling = 1e+09
)
}
\arguments{
\item{count.matrix}{a count matrix by mapping reads from multiple samples to a pangenome of one species, each row is one gene and each column is one sample}

\item{pangenome.file}{the pangenome.csv file generated by panphlan}

\item{min.cov}{minimum rpkm for the gene family (default: 10)}

\item{frac.gene}{minimum fraction of gene families has count > min.cov relative to the minimum gene family number (default: 1)}

\item{scaling}{a scaling factor to normalize the library size for each sample (default: 1e9)}
}
\description{
filter input data
}
\author{
Xbiome
}
