% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.strain.sample.r
\name{solveStrainSampleRegression}
\alias{solveStrainSampleRegression}
\title{solveStrainSampleRegression}
\usage{
solveStrainSampleRegression(D, P, ncpu = 1, fil.thre = 2, ...)
}
\arguments{
\item{D}{a normalized count matrix (output of `preprocess`)}

\item{P}{current gene-strain matrix}

\item{ncpu}{number of processes running in parallel (default: 1)}

\item{fil.thre}{threshold for filtering abnormal gene families (default: 2)}

\item{...}{other parameters for nnlm}
}
\description{
solveStrainSampleRegression
}
\author{
Xbiome
}
