% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.r
\name{strain.decompose}
\alias{strain.decompose}
\title{strain.decompose}
\usage{
strain.decompose(
  obj,
  rank = NULL,
  jc.threshold = 0.1,
  abun.threshold = 0.1,
  max.strain = 12,
  presence.thre = 0.5,
  ncpu = 1,
  seed = 123456,
  debug = FALSE
)
}
\arguments{
\item{obj}{strainpandar object (output of `preprocess`)}

\item{rank}{the rank of NMF (number of strains)}

\item{presence.thre}{in iterative process, for a gene to be considered as present, this threshold is used as `adjsut` for the function `gene.presence` (default: 0.5)}

\item{ncpu}{number of processes running in parallel (default: 1)}

\item{seed}{random seed for NMF}

\item{debug}{print debugging information (default: FALSE)}
}
\description{
Decovolute strain information based on pangenome coverage information
}
\author{
Xbiome
}
