% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.r
\name{strainpanda.init}
\alias{strainpanda.init}
\title{strainpanda.init}
\usage{
strainpanda.init(
  obj,
  ncpu = 1,
  abundance.thre = 0.05,
  fil.thre = 2,
  cutree.h = 0.1
)
}
\arguments{
\item{obj}{strainpandar object generated using preprocess}

\item{ncpu}{number of processes running in parallel (default: 1)}

\item{abundance.thre}{threshold for relative abundance (default: 0.05)}

\item{fil.thre}{threshold for filtering abnormal gene families (default: 2)}

\item{cutree.h}{the height at which the clustering tree is cut (defualt: 0.1)}
}
\description{
select most possible genomes from the reference database
}
\author{
Xbiome
}
